Convert :add-package to AddPackageCommand
Lift the :add-package handler into a standalone IFalloutCommand type. The shared
AddOrReplacePackage helper and PACKAGE_TYPE_*/BUILD_PROJECT_FILE constants stay on
Program as residual (still used by cake and tests) until the #392 collapse PR moves
them into services. Replaces the DelegateCommand adapter.
